+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -7,6 +7,10 @@
   ([#1524](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1524))
 * The `OtlpExporter` class has been made internal.
   ([#1528](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1528))
+* Removed `ServiceName` from options available on the `AddOtlpExporter`
+  extension. It is not required by the
+  [specification](https://github.com/open-telemetry/opentelemetry-specification/blob/master/specification/protocol/exporter.md#configuration-options).
+  ([#1557](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1557))
 
 ## 0.8.0-beta.1